clients/js: re-org sui code
