+++ clients/js/sui/index.ts
@@ -1,0 +1,3 @@
+export * from "./publish";
+export * from "./types";
+export * from "./utils";
